BIKE Round-3 runtime code path selection based on CPU capabilities
This commit imports the new code from the BIKE repository
https://github.com/awslabs/bike-kem
that implements runtime code path selection based on the CPU
capabilities. This allows the x86_64 optimized code of BIKE
to be included in s2n and significantly improve BIKE performance.

Call-out: the CMake build system is updated to properly
compile the optimized files. However, the Makefile system
still compiles only the portable code and should be updated
in a follow-up commit.

Signed-off-by: Dusan Kostic <dkostic@amazon.com>